backend/pyproject.toml
[project]
name = "recruiter-agent-backend"
version = "0.1.0"
description = "Voice recruitment screening agent using Twilio"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "websockets>=11.0.0",
    "twilio>=8.10.0",
    "python-multipart>=0.0.6",
]
